requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
markdown==3.7
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.12
mcp[cli]==1.9.2
